refactor(backend): change name of migration
